Implement MongoDB settings and user settings management; refactor related classes and interfaces (#5)

+++ CoBySi.Pomodoro/Models/PomodoroSettings.cs
@@ -2,8 +2,8 @@ namespace CoBySi.Pomodoro;
 
 public class PomodoroSettings
 {
-    public int MinutesPerPomodoro { get; set; }
-    public int MinutesPerShortBreak { get; set; }
-    public int MinutesPerLongBreak { get; set; }
-    public int PomodorosBeforeLongBreak { get; set; }
+    public double MinutesPerPomodoro { get; set; }
+    public double MinutesPerShortBreak { get; set; }
+    public double MinutesPerLongBreak { get; set; }
+    public double PomodorosBeforeLongBreak { get; set; }
 }
